requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
get==2019.4.13
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
post==2019.4.13
psycopg2-binary==2.8.4
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
SQLAlchemy==1.3.12
Werkzeug==0.16.0
